+++ lib/src/model/transform/tile_translate.dart
@@ -4,14 +4,15 @@ import '../tile_data_model.dart';
 
 class TileTranslate {
   final TilePoint _offset;
+  final double scale;
 
-  TileTranslate(this._offset);
+  TileTranslate(this._offset, {this.scale = 1.0});
 
   TileData translate(TileData tile) => TileData(
       layers: tile.layers.map(_translateLayer).toList(growable: false));
 
   TileDataLayer _translateLayer(TileDataLayer layer) {
-    final pixelsPerTileUnit = 1 / layer.extent * tileSize;
+    final pixelsPerTileUnit = 1.0 / layer.extent * tileSize;
     final translation =
         TilePoint(_offset.x / pixelsPerTileUnit, _offset.y / pixelsPerTileUnit);
     return TileDataLayer(
@@ -39,7 +40,7 @@ class TileTranslate {
               : null);
 
   TilePoint _translatePoint(TilePoint point, TilePoint translation) =>
-      TilePoint(point.x + translation.x, point.y + translation.y);
+      _scale(TilePoint(point.x + translation.x, point.y + translation.y));
 
   TileLine _translateLine(TileLine line, TilePoint translation) =>
       TileLine(line.points
@@ -50,4 +51,11 @@ class TileTranslate {
       TilePolygon(polygon.rings
           .map((l) => _translateLine(l, translation))
           .toList(growable: false));
+
+  TilePoint _scale(TilePoint point) {
+    if (scale == 1.0) {
+      return point;
+    }
+    return point * scale;
+  }
 }