+++ lib/src/themes/theme_reader.dart
@@ -124,10 +124,14 @@ class ThemeReader {
     final paintJson = jsonLayer['paint'];
     final opacity = expressionParser.parse(paintJson?['raster-opacity'],
         whenNull: () => LiteralExpression(1.0));
+    final resampling = expressionParser.parse(paintJson?['raster-resampling'],
+        whenNull: () => LiteralExpression("linear"));
     return ThemeLayerRaster(
         jsonLayer['id'] ?? _unknownId, ThemeLayerType.raster,
         selector: selector,
-        paintModel: RasterPaintModel(opacity: opacity.asDoubleExpression()),
+        paintModel: RasterPaintModel(
+            opacity: opacity.asDoubleExpression(),
+            rasterResampling: resampling.asOptionalStringExpression()),
         minzoom: _minZoom(jsonLayer),
         maxzoom: _maxZoom(jsonLayer),
         metadata: _metadata(jsonLayer));
@@ -256,7 +260,7 @@ class ThemeReader {
     final font = layout?['text-font'];
     String? fontFamily;
     FontStyle? fontStyle;
-    if (font is List<dynamic>) {
+    if (font is List<dynamic> && font.isNotEmpty) {
       fontFamily = font[0];
       if (fontFamily != null && fontFamily.toLowerCase().contains("italic")) {
         fontStyle = FontStyle.italic;